Move function to reduce duplication
