+++ src/OpenLoco/src/Viewport.cpp
@@ -54,96 +54,6 @@ namespace OpenLoco::Ui
         paint(drawingCtx, screenToViewport(intersection));
     }
 
-    struct StationBorder
-    {
-        uint32_t left;
-        uint32_t right;
-        uint16_t width;
-    };
-    static constexpr std::array<StationBorder, 4> kZoomToStationBorder = {
-        StationBorder{
-            ImageIds::curved_border_left_medium,
-            ImageIds::curved_border_right_medium,
-            3,
-        },
-        StationBorder{
-            ImageIds::curved_border_left_medium,
-            ImageIds::curved_border_right_medium,
-            3,
-        },
-        StationBorder{
-            ImageIds::curved_border_left_small,
-            ImageIds::curved_border_right_small,
-            1,
-        },
-        StationBorder{
-            ImageIds::curved_border_left_small,
-            ImageIds::curved_border_right_small,
-            1,
-        },
-    };
-
-    static constexpr std::array<Gfx::Font, 4> kZoomToStationFonts = {
-        Gfx::Font::medium_bold,
-        Gfx::Font::medium_bold,
-        Gfx::Font::small,
-        Gfx::Font::small,
-    };
-
-    // 0x0048DF4D, 0x0048E13B
-    static void drawStationName(Gfx::DrawingContext& drawingCtx, const Station& station, uint8_t zoom, bool isHovered)
-    {
-        const Gfx::RenderTarget& unZoomedRt = drawingCtx.currentRenderTarget();
-        if (!station.labelFrame.contains(unZoomedRt.getDrawableRect(), zoom))
-        {
-            return;
-        }
-
-        auto& borderImages = kZoomToStationBorder[zoom];
-
-        const auto companyColour = [&station]() {
-            if (station.owner == CompanyId::null)
-            {
-                return Colour::grey;
-            }
-            else
-            {
-                return CompanyManager::getCompanyColour(station.owner);
-            }
-        }();
-        const auto colour = Colours::getTranslucent(companyColour, isHovered ? 0 : 1);
-
-        Ui::Point topLeft = { station.labelFrame.left[zoom],
-                              station.labelFrame.top[zoom] };
-        Ui::Point bottomRight = { station.labelFrame.right[zoom],
-                                  station.labelFrame.bottom[zoom] };
-
-        drawingCtx.drawImage(topLeft, ImageId(borderImages.left).withTranslucency(ExtColour::unk34));
-        drawingCtx.drawImage(topLeft, ImageId(borderImages.left).withTranslucency(colour));
-
-        Ui::Point topRight = { static_cast<int16_t>(bottomRight.x - borderImages.width), topLeft.y };
-        drawingCtx.drawImage(topRight, ImageId(borderImages.right).withTranslucency(ExtColour::unk34));
-        drawingCtx.drawImage(topRight, ImageId(borderImages.right).withTranslucency(colour));
-
-        drawingCtx.drawRect(topLeft.x + borderImages.width + 1, topLeft.y, bottomRight.x - topLeft.x - 2 * borderImages.width, bottomRight.y - topLeft.y + 1, enumValue(ExtColour::unk34), Gfx::RectFlags::transparent);
-        drawingCtx.drawRect(topLeft.x + borderImages.width + 1, topLeft.y, bottomRight.x - topLeft.x - 2 * borderImages.width, bottomRight.y - topLeft.y + 1, enumValue(colour), Gfx::RectFlags::transparent);
-
-        char buffer[512]{};
-
-        FormatArguments args;
-        args.push<uint16_t>(enumValue(station.town));
-        auto* str = buffer;
-        *str++ = ControlCodes::Colour::black;
-        str = StringManager::formatString(str, station.name, args);
-        *str++ = ' ';
-        StringManager::formatString(str, getTransportIconsFromStationFlags(station.flags));
-
-        auto tr = Gfx::TextRenderer(drawingCtx);
-        tr.setCurrentFont(kZoomToStationFonts[zoom]);
-        auto point = topLeft + Point(borderImages.width, 0);
-        tr.drawString(point, Colour::black, buffer);
-    }
-
     // 0x0048DE97
     static void drawStationNames(Gfx::DrawingContext& drawingCtx)
     {
